# Update API and add performance enhancements to CollectionTables