--- a/setup.py
+++ b/setup.py
@@ -37,6 +37,7 @@ setuptools.setup(
         'pytz>=2018.4',
         'tzlocal>=1.5.1',
         'pandas>=1.0.0',
+        'tqdm',
     ),
     python_requires='>=3.6',
     extras_require={