--- a/src/resdk/__init__.py
+++ b/src/resdk/__init__.py
@@ -1,3 +1,4 @@
 """Resolwe SDK for Python."""
+from .collection_tables import CollectionTables  # noqa
 from .resdk_logger import log_to_stdout, start_logging  # noqa
 from .resolwe import Resolwe, ResolweQuery  # noqa