--- a/src/resdk/utils/table_cache.py
+++ b/src/resdk/utils/table_cache.py
@@ -1,16 +1,18 @@
 """Cache util functions for CollectionTables."""
 import os
+import pickle
 import sys
-from datetime import datetime
-from typing import Optional
-
-import pandas as pd
+from shutil import rmtree
+from typing import Any
 
 from resdk.__about__ import __version__
 
 
-def default_cache_dir() -> str:
-    """Return default cache directory specific for the current OS. Code from Orange3.misc.environ."""
+def _default_cache_dir() -> str:
+    """Return default cache directory specific for the current OS.
+
+    Code originally from Orange3.misc.environ.
+    """
     if sys.platform == "darwin":
         base = os.path.expanduser("~/Library/Caches")
     elif sys.platform == "win32":
@@ -19,8 +21,22 @@ def default_cache_dir() -> str:
         base = os.getenv("XDG_CACHE_HOME", os.path.expanduser("~/.cache"))
     else:
         base = os.path.expanduser("~/.cache")
+    return base
 
-    base = os.path.join(base, "ReSDK", __version__)
+
+def cache_dir_resdk_base() -> str:
+    """Return base ReSDK cache directory."""
+    return os.path.join(_default_cache_dir(), "ReSDK")
+
+
+def cache_dir_resdk() -> str:
+    """Return ReSDK cache directory."""
+    v = __version__
+    if "dev" in v:
+        # remove git commit hash
+        v = v[: v.find("dev") + 3]
+    base = os.path.join(cache_dir_resdk_base(), v)
+
     if sys.platform == "win32":
         # On Windows cache and data dir are the same.
         # Microsoft suggest using a Cache subdirectory
@@ -29,32 +45,30 @@ def default_cache_dir() -> str:
         return base
 
 
-def load_cache(
-    directory: str, collection_slug: str, data_type: str, modified: datetime
-) -> Optional[pd.DataFrame]:
-    """If there is a cached file for the same data return it."""
-    full_cache_file = _full_cache_file(directory, collection_slug, data_type, modified)
-    if os.path.exists(full_cache_file):
-        return pd.read_pickle(full_cache_file)
+def clear_cache_dir_resdk() -> None:
+    """Delete all cache files from the default cache directory."""
+    cache_dir = cache_dir_resdk_base()
+    if os.path.exists(cache_dir):
+        rmtree(cache_dir)
 
 
-def save_cache(
-    df: pd.DataFrame,
-    directory: str,
-    collection_slug: str,
-    data_type: str,
-    modified: datetime,
-):
-    """If not existing, save data to cache file."""
-    full_cache_file = _full_cache_file(directory, collection_slug, data_type, modified)
-    if not os.path.exists(full_cache_file):
-        df.to_pickle(full_cache_file)
+def load_pickle(pickle_file: str) -> Any:
+    """Load object from the pickle file.
+
+    :param pickle_file: file path
+    :return: un-pickled object
+    """
+    if os.path.exists(pickle_file):
+        return pickle.load(open(pickle_file, "rb"))
 
 
-def _full_cache_file(
-    directory: str, collection_slug: str, data_type: str, modified: datetime
-) -> str:
-    """Return full path to specified cache file."""
-    datetime_str = modified.isoformat()
-    cache_file = f"{collection_slug}_{data_type}_{datetime_str}.pickle"
-    return os.path.join(directory, cache_file)
+def save_pickle(obj: Any, pickle_file: str, override=False) -> None:
+    """Save given object into a pickle file.
+
+    :param obj: object to bi pickled
+    :param pickle_file: file path
+    :param override: if True than override existing file
+    :return:
+    """
+    if not os.path.exists(pickle_file) or override:
+        pickle.dump(obj, open(pickle_file, "wb"))