… tests/unit/test_collection_tables.py
… tests/unit/test_tables_cache.py