What started with just a few fixes ended up being closer to a rewrite.
The major changes are:
 * change main methods to attributes: `metadata()` -> `meta`, `expressions()` -> `exp`, `counts()` -> `rc`
 * add `clear_cache()` which removes all cache files from the default location
 * add method `id_to_symbol` which holds gene id to symbole mapping
 * `exp` and `rc` return gene ids instead of symboles
 * sample name is used instead of sample slug for tables index
 * add separate versioning for expression data
 * seperatly fetch metadata and expressions
 * add downloading progress bar
 * mapping of gene ids to symboles is now cached on disc and is also used on other collection
 * add `CollectionTable` to `resdk` namespace
 * store expression type in returned tables `attrs['exp_type']` attribute

Performance on 45 samples (`exp` and `rc` have the same performance):
| how  | what | old | new |
|--------|--------|------|-------|
| / | init | 0.4s | <0.01s |
| uncached | meta | 127s* | 1.7s |
| uncached | exp | 255s* | 64s** |
| uncached | id_to_symbol | / | 86s |
| cached | meta | 1.7s | 0.4s |
| cached | exp | 0.02s | 1.8s |
| cached | id_to_symbol | / | 2.8s |

\* calling just one of `metadata()` or `expressions()`, if used together resolwe server gets queried only once and performance improves for he second function for ~120s
** without gene id to symbol mapping