@JureZmrzlikar 2020-12-07
> @JureZmrzlikar Can you look at Jenkins and help me understand why it is failing?

I have no clue. @acopar Do you have any idea?