@acopar 2020-12-09
This problem happened because pip changed switched the default dependency resolver on the recent 20.3.* version. The new resolver has different dependency resolution, and causes some glitches with dependencies, in particular backtracking package versions until timeout. This PR should fix this problem:
https://github.com/genialis/genialis-base/pull/596

We can re-trigger build after genialis-base PR is merged.